requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flask-Mail==0.9.1
Flask-Sock==0.7.0
email-validator==2.0.0
Pillow==10.2.0
python-dotenv==1.0.1
Werkzeug==3.0.1 
exifread==3.0.0
flask-sock==0.7.0
gunicorn==21.2.0 